There's a logic keyboard now. Integrals and normal.
I originally intded for the shift to be numbers, and operators (couldn't find enough so there are some letters still)
The Shift also has a problem in all of them that it does not have the option to stay (you input one character and it returns to main).
As for the Symbols in the logic and Integrals keyboard, they are yet to be changed. I intended for them to have the same one. And for it to have the  subscripts and superscripts of all letters and numbers (including capital) but I was unable.
Engineering Keyboard could be used as an example for the menu.